Move resolved models to PackageGraph module
Extract ResolvedPackage into separate file

Extract ResolvedProduct into separate file

Extract ResolvedTarget into separate file

Reorganize protocol conformance
