+++ Tests/PackageGraphTests/TargetTests.swift
@@ -11,6 +11,7 @@
 import XCTest
 import TSCBasic
 
+import PackageGraph
 import PackageModel
 
 private extension ResolvedTarget {
